Move out schema source tests
